split server_test.go into domain-specific test files
Extract auth/JWE, client, dynamic tools, query, resources, and settings
tests into separate files for maintainability; remove isSelectQuery from
server_query.go as it moved to the query test package.
